+++ tests/testthat/test-cnd-abort.R
@@ -452,7 +452,7 @@ test_that("headers and body are stored in respective fields", {
   local_use_cli()  # Just to be explicit
 
   cnd <- catch_cnd(abort(c("a", "b", i = "c")), "error")
-  expect_equal(cnd$message, "a")
+  expect_equal(cnd$message, set_names("a", ""))
   expect_equal(cnd$body, c("b", i = "c"))
 })